+++ js/firebase/user.js
@@ -1,19 +1,13 @@
 import { HOURS as HOUR } from '@shgysk8zer0/consts/date.js';
-import { deleteUser as dUser, updateProfile as uProfile } from 'firebase/firebase-auth.js';
-
-import { getCurrentUser } from './auth.js';
-
-export async function updateProfile({ name: displayName, image: photoURL }) {
-	const user = await getCurrentUser();
-
-	if (user) {
-		await uProfile(user, { displayName, photoURL });
-	}
-}
+import { getCurrentUser } from '@aegisjsproject/firebase-account-routes/auth.js';
+import { ORG_TOKEN_KEY } from '../consts.js';
 
 export async function getToken() {
 	const user = await getCurrentUser();
-	return user.getIdToken();
+
+	if (! user.isAnonymous) {
+		return user.getIdToken();
+	}
 }
 
 export async function createTokenCookie(name = 'token') {
@@ -21,26 +15,33 @@ export async function createTokenCookie(name = 'token') {
 	await cookieStore.set({ name, value, expires: Date.now() + HOUR, sameSite: 'strict' });
 }
 
-export async function deleteUser() {
-	const user = await getCurrentUser();
+export const hasOrgToken = () => localStorage.hasOwnProperty(ORG_TOKEN_KEY);
 
-	if (user) {
-		await dUser(user);
-	}
-}
+export const hasValidOrgToken = () => hasOrgToken() && (parseInt(localStorage.getItem(ORG_TOKEN_KEY)) || -1) > Date.now();
 
 export async function genOrgToken({ signal } = {}) {
-	const token = await getToken();
-
-	if (typeof token === 'string') {
-		const resp = await fetch('/api/orgJWT', {
-			headers: { Authorization: `Bearer ${token}` },
-			signal,
-			credentials: 'same-origin',
-		});
-
-		return resp.ok ? await resp.json() : null;
+	if (hasValidOrgToken()) {
+		return Math.max(parseInt(localStorage.getItem(ORG_TOKEN_KEY), 0)) || -1;
 	} else {
-		return null;
+		const token = await getToken();
+
+		if (typeof token === 'string') {
+			const resp = await fetch('/api/orgJWT', {
+				headers: { Authorization: `Bearer ${token}` },
+				signal,
+				credentials: 'same-origin',
+			});
+
+			if (resp.ok) {
+				const { expires } = await resp.json();
+				localStorage.setItem(ORG_TOKEN_KEY, expires);
+				return expires;
+			} else {
+				return -1;
+			}
+		} else {
+			return -1;
+		}
 	}
+
 }