+++ rollup.config.js
@@ -14,11 +14,14 @@ export default {
 		'@aegisjsproject/router',
 		'@aegisjsproject/state/state.js',
 		'@aegisjsproject/state',
+		'@aegisjsproject/firebase-account-routes',
+		'@aegisjsproject/firebase-account-routes/auth.js',
 	],
 	output: {
 		file: 'js/index.min.js',
 		format: 'esm',
 		plugins: [terser()],
 		sourcemap: true,
+		inlineDynamicImports: true,
 	}
 };